feat(core): count a magnitude and a bound, and leave the unit to the application
A level is a sum of movements and nothing in this engine multiplies or divides
one, so a binary float bought nothing and cost associativity. Measured, in the
predicate that decides `OutOfBounds`, with the expression `resource.rs` uses:

    0.1 + 0.2 + 0.3  ->  0.60000000000000008882    outside [0, 0.6]
    0.3 + 0.2 + 0.1  ->  0.59999999999999997780    inside

The two hypotheses a caller may ask under fold in different orders — one by
CommitmentId, one by due date with a running total — so that is not a rounding
nuisance. It is a way for two answers over identical knowledge to disagree with
each other for no reason in the domain, and which one you get depends on a
sha256 ordering.

A bound is now `i128` and a magnitude is `u128`. Two types rather than one
because they are two things: a magnitude's direction is the statement's to give,
and a level is a sum of both directions and can be negative. What the count
counts is **not** in the engine. Cents, whole items, thirds of an hour, pallets
of forty-eight — the engine adds and compares, and naming the unit here, as a
decimal scale or as anything else, would make every application share one
approach to units. A scale was drafted and removed for exactly that reason: it
reproduced, one layer up, the industrial premise that had chosen the float.

The core comes out smaller than it went in. `NonFinite` is gone from bounds and
from magnitudes; six of `Constraint`'s seven constructors cannot fail and no
longer pretend to; `check` is a comparison again; `ConstraintError` has one
variant where it had two; and `ActionValueError` has one where it had two, since
a negative magnitude stopped being writable rather than being caught. That test
shrank, which is what moving an invariant into a type looks like.

Three refusals are new and each is the engine's own arithmetic, where recusing
is the only outcome that does not invent a number:

  ActionValueError::Zero          the one value the type admits and the ontology
                                 does not
  MagnitudeUnmovable             an unsigned magnitude too large to become a
                                 signed movement — the whole price of the
                                 unsigned side, paid where the sign enters
  LevelOutOfRange                movements that sum past what a count holds; a
                                 wrapped total would be weighed against a bound
                                 as though it were a level

The last two had no test and would have been guards nobody had seen answer.
Both are exercised, and all three were mutated and watched go red.

Two identities moved, and only two: `Resource` through its `Constraint`, and
`Commitment` through its `ActionValue`. The pin says so, and its docstring now
says why that is not the blast radius on a repository — the pins use synthetic
references, and in a real history one moved `Resource` moves everything
downstream of it.

The application's own decision, forced by this one and recorded where it is
made: a count is written as a decimal string. Serde cannot read a 128-bit
integer out of an internally tagged enum, and every record here is tagged so the
file says what each entry is; and a JSON *number* is a double to most things
that read JSON, so the exactness the engine gained would have stopped at the
file. It is attached to the fields rather than to a newtype, so nothing that
merely builds a record has to know how one is written.

The arrangements did not need to change, and that is the measurement worth
keeping: every value in both laboratory rows was already whole. `00-reconstruction`
says why, in the subject, and has since the first experiment —

    "Feasibility accumulates levels in f64, where addition is not associative …
     Integers keep the experiment measuring reconstruction rather than float
     determinism."

— and `agents/world.rs` says the same thing independently. The hazard was found,
written down and worked around by discipline in every arrangement of both rows,
for seven experiments, and never raised to the engine. Both notes now record that
it is gone from underneath them. No published number moved: 100.0 became 100.

Twelve of the files in this diff are not authored. The four recorded repositories
in `agents/04-multiagent` were re-derived by a throwaway instrument: identities as
written came from a dump taken in a detached worktree at this commit's parent, and
each was paired positionally with the identity derived in its place — a journal is
a sequence, and the nth admission is the nth admission. It rebuilt each repository
whole in memory before writing any of it, which is what `converge` already
promises a party. Verified against what was written: the same magnitudes in the
same order in all four, and the same counts of entries, decisions and worlds. The
instrument is not kept; the file as its author wrote it is in this commit's parent.

One residue, and it is not fixable: the run directories' README and ANSWER quote
identities in prose, and those no longer resolve. Migrating them would be
rewriting what an agent was handed and what it answered, which is the one thing
those files exist to hold. They stay as written, and an identity quoted in prose
turns out to be the one reference a migration cannot follow.

+++ cli/src/reading.rs
@@ -73,7 +73,8 @@ pub enum ConflictRecord {
     },
     OutOfBounds {
         instance: String,
-        level: f64,
+        #[serde(with = "crate::journal::count")]
+        level: i128,
     },
 }
 
@@ -162,7 +163,13 @@ pub struct Reading {
     pub frozen: BTreeSet<String>,
     pub open: BTreeSet<String>,
     pub conditions: BTreeMap<String, ConditionRecord>,
-    pub level: f64,
+    /// What has settled on the instance asked about.
+    ///
+    /// Written as a decimal string, for the reason [`crate::journal::count`] gives: a reading crosses
+    /// a process boundary, and that is where a JSON number stops being exact for anyone who is not
+    /// this program.
+    #[serde(with = "crate::journal::count")]
+    pub level: i128,
     pub conflicts: Vec<ConflictRecord>,
 }
 
@@ -411,8 +418,8 @@ mod tests {
             Admission::Resource {
                 label: "cash".into(),
                 kind: ResourceKindRecord::Between {
-                    lower: 0.0,
-                    upper: 100.0,
+                    lower: 0,
+                    upper: 100,
                 },
                 recorded_at: day(1),
             },
@@ -465,7 +472,7 @@ mod tests {
         // Two commitments, and an Event settling one of them. Both halves of the partition are
         // therefore non-empty, which is what makes the assertions below bite: a world with nothing
         // frozen and nothing open would have this guard comparing empty sets and reporting agreement.
-        let flowing = |magnitude: f64| Admission::Commitment {
+        let flowing = |magnitude: u128| Admission::Commitment {
             accountable: one,
             executors: [one].into(),
             beneficiaries: [other].into(),
@@ -478,7 +485,7 @@ mod tests {
             recorded_at: day(2),
         };
 
-        journal.extend([flowing(10.0), flowing(20.0)]);
+        journal.extend([flowing(10), flowing(20)]);
         crate::journal::replay_remaining(&mut canon, &journal, &mut admitted).expect("admissible");
 
         let (settled, open) = (admitted.commitments[0], admitted.commitments[1]);
